+++ packages/analyzer_kit_annotation/lib/src/serialize.dart
@@ -1,5 +1,0 @@
-final class Serialize {
-  const Serialize();
-}
-
-const serialize = Serialize();